requirements.txt
# OpenAI Agents SDK
openai>=1.0.0
openai-agents-sdk>=0.1.0

# Map and geocoding services
geopy>=2.4.0
folium>=0.15.0
requests>=2.31.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
